requirements.txt
PyMySQL==1.0.2
PyYAML==6.0
sphinx
nose
pyinstaller==5.6.2